requirements/base.txt
Django==1.8
django-admin-bootstrapped==2.5.7
django-authtools==1.4.0
django-braces==1.8.1
django-crispy-forms==1.6.0
django-environ==0.4.0
easy-thumbnails==2.3
django-markdownx==1.4.2
django-rest-swagger==0.3.5
djangorestframework==3.3.3